legal-workbench/poc-fasthtml-stj/requirements.txt
python-fasthtml>=0.4.0
httpx>=0.27.0
python-multipart>=0.0.9
uvicorn>=0.30.0
